fix: int64 indexing migration batch 6 - test fixes for NumPy 2.x int64 default
Root cause identified: "memory corruption" errors were NOT actual corruption.
Tests were calling GetInt32() on Int64 arrays (np.arange now returns int64).

Fixes:
- np.random.shuffle.cs: NextInt64 → NextLong (correct method name)
- np.random.shuffle.cs: SwapSlicesAxis0 int → long parameters
- BattleProofTests.cs: GetInt32 → GetInt64 for arange-based tests
- np.transpose.Test.cs: long[] → int[] for axis array (axes stay int)
- ReadmeExample.cs: cast n_samples to int for np.ones() calls
- NpApiOverloadTests: int → long for count_nonzero return, NDArray<int>[] → NDArray<long>[] for nonzero
- BooleanIndexing.BattleTests.cs: shape.SequenceEqual(new[]) → shape.SequenceEqual(new long[])
- Updated INT64_MIGRATION_PROGRESS.md with root cause analysis

+++ test/NumSharp.UnitTest/Backends/Kernels/BattleProofTests.cs
@@ -26,8 +26,8 @@ public class BattleProofTests : TestClass
 
         var z = y * 2;
 
-        Assert.AreEqual(2, z.GetInt32(0), "y[0]*2 = 1*2 = 2");
-        Assert.AreEqual(6, z.GetInt32(1), "y[1]*2 = 3*2 = 6");
+        Assert.AreEqual(2L, z.GetInt64(0), "y[0]*2 = 1*2 = 2");
+        Assert.AreEqual(6L, z.GetInt64(1), "y[1]*2 = 3*2 = 6");
     }
 
     [Test]
